Extract self-build connect routes out of submissions.ts
Moves GET /api/submissions/:id/connect (mints the creator-wide agent key +
install snippets for a self-build round) and the two retired per-game
agent-key routes into a new apps/api/src/agent-surface/self-build-connect-routes.ts
module, alongside the existing self-build-connect.ts payload templates.

ensureSubmissionSlug stays in submissions.ts (shared with createGame via
confirmSlugClaim/isSlugClaimed) and is passed in as an injected option, same
pattern as checkUserAccess.

submissions.ts drops from 4628 to 4497 lines; the new file is 159 lines,
0 comment-prose debt.
